refactor: GameResult의 상수 활용 및 패키지 이동
