+++ src/test/java/blackjack/domain/game/StatusTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card;
+package blackjack.domain.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
